make inverse work

+++ Blatt2/Matrix/CMatrix/CMatrix.h
@@ -8,14 +8,14 @@ class CMatrix {
 	private:
 	unsigned short int i, j, k;  // for loops
 	int rows, cols;
-	int **matrix;
+	float **matrix;
 
 	public:
-	CMatrix(int **values, const unsigned short int dim) : 
+	CMatrix(float **values, const unsigned short int dim) : 
 		CMatrix(values, dim, dim) {}
 	CMatrix(const unsigned short int rowDim, const unsigned short int colDim) :
 		CMatrix(nullptr, rowDim, colDim) {}
-	CMatrix(int **values, const unsigned short int rowDim, const unsigned short int colDim);
+	CMatrix(float **values, const unsigned short int rowDim, const unsigned short int colDim);
 	CMatrix(const CMatrix &values) : 
 		CMatrix(nullptr, values.getRows(), values.getCols()) {this->matSum(&values);}
 
@@ -23,11 +23,11 @@ class CMatrix {
 	void matMult(const CMatrix *mat);
 	void transpose();
 	void inverse();
-	long int scalarProduct(const CMatrix *mat); // multiplies, then sums elements at same locations
+	float scalarProduct(const CMatrix *mat); // multiplies, then sums elements at same locations
 
 	unsigned short int getRows() const;
 	unsigned short int getCols() const;
-	int getElem(unsigned short int n, unsigned short int m) const; // start counting at 0
+	float getElem(unsigned short int n, unsigned short int m) const; // start counting at 0
 
 	void print();
